updating nonce and amount columns to handle uint256
- the bigints are signed so they can't handle the largest positive values of
uint256 but I'll add support for that only if we think its necessary.

+++ lndr-backend/db/create_tables.sql
@@ -1,20 +1,20 @@
 CREATE TABLE pending_credits (
     hash        CHAR(64) PRIMARY KEY,
     submitter   CHAR(40),
-    nonce       INT,
+    nonce       BIGINT,
     creditor    CHAR(40),
     debtor      CHAR(40),
-    amount      INT,
+    amount      BIGINT,
     memo        CHAR(32),
     signature   CHAR(130)
 );
 
 CREATE TABLE verified_credits (
     hash               CHAR(64) PRIMARY KEY,
-    nonce              INT,
+    nonce              BIGINT,
     creditor           CHAR(40),
     debtor             CHAR(40),
-    amount             INT,
+    amount             BIGINT,
     memo               CHAR(32),
     creditor_signature CHAR(130),
     debtor_signature   CHAR(130)
